+++ src/csaf_lib/models/__init__.py
@@ -1,15 +1,15 @@
 """Data models for CSAF VEX."""
 
-from csaf_vex.models.common import Acknowledgment, Note, Reference
-from csaf_vex.models.csafvex import CSAFVEX
-from csaf_vex.models.document import Document
-from csaf_vex.models.product_tree import (
+from csaf_lib.models.common import Acknowledgment, Note, Reference
+from csaf_lib.models.csafvex import CSAFVEX
+from csaf_lib.models.document import Document
+from csaf_lib.models.product_tree import (
     Branch,
     FullProductName,
     ProductTree,
     Relationship,
 )
-from csaf_vex.models.vulnerability import (
+from csaf_lib.models.vulnerability import (
     CWE,
     ID,
     Flag,